kafka/requirements.txt
kafka-python==2.3.1
snowflake-connector-python==3.12.2
